moved PathResolver in dedicated module, and updated references.
refs #704
