+++ gnrpy/gnr/dev/cli/gnrmkproject.py
@@ -9,7 +9,8 @@ usage: gnrmkproject projectname
 import os
 
 from gnr.core.cli import GnrCliArgParse
-from gnr.app.gnrdeploy import ProjectMaker, InstanceMaker, PackageMaker, PathResolver
+from gnr.app.gnrdeploy import ProjectMaker, InstanceMaker, PackageMaker
+from gnr.app.pathresolver import Pathresolver
 
 description = "Bootstrap a new project folder and subfolders"
